@@ -11,27 +11,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^15.2.7",
-    "@angular/animations": "^15.2.8",
-    "@angular/cdk": "^14.2.6",
-    "@angular/cli": "^15.2.7",
-    "@angular/common": "^15.2.8",
-    "@angular/compiler": "^15.2.8",
-    "@angular/compiler-cli": "^15.2.8",
-    "@angular/core": "^15.2.8",
-    "@angular/forms": "^15.2.8",
-    "@angular/material": "^14.2.6",
-    "@angular/material-moment-adapter": "^14.2.6",
-    "@angular/platform-browser": "^15.2.8",
-    "@angular/platform-browser-dynamic": "^15.2.8",
-    "@angular/router": "^15.2.8",
+    "@angular-devkit/build-angular": "^17.0.9",
+    "@angular/animations": "^17.0.8",
+    "@angular/cdk": "^17.1.1",
+    "@angular/cli": "^17.0.9",
+    "@angular/common": "^17.0.8",
+    "@angular/compiler": "^17.0.8",
+    "@angular/compiler-cli": "^17.0.8",
+    "@angular/core": "^17.0.8",
+    "@angular/forms": "^17.0.8",
+    "@angular/material": "^17.1.1",
+    "@angular/material-moment-adapter": "^17.1.1",
+    "@angular/platform-browser": "^17.0.8",
+    "@angular/platform-browser-dynamic": "^17.0.8",
+    "@angular/router": "^17.0.8",
     "@babel/core": "^7.14.2",
     "@formatjs/intl-displaynames": "^6.2.6",
     "@fortawesome/angular-fontawesome": "^0.10.2",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@ngrx/effects": "^15.4.0",
-    "@ngrx/store": "^15.4.0",
+    "@ngrx/effects": "^17.0.1",
+    "@ngrx/store": "^17.0.1",
     "@ngx-translate/core": "^14.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "@turf/boolean-point-in-polygon": "^6.5.0",
@@ -62,10 +62,10 @@
     "moment": "^2.29.4",
     "moment-range": "^4.0.2",
     "ngx-clipboard": "^15.0.0",
-    "ngx-cookie-service": "^15.0.0",
+    "ngx-cookie-service": "^17.0.1",
     "ngx-cookieconsent": "^4.0.1",
-    "ngx-csv-parser": "^1.2.2",
-    "ngx-toastr": "^16.1.1",
+    "ngx-csv-parser": "^1.3.1",
+    "ngx-toastr": "^17.0.2",
     "nouislider": "^15.6.1",
     "npm": "^8.11.0",
     "ol": "^7.4.0",
@@ -81,29 +81,29 @@
     "timers": "^0.1.1",
     "tippy.js": "^5.2.1",
     "tslib": "^2.2.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.2.2",
     "ua-parser-js": "^0.7.28",
     "uuid": "^7.0.3",
     "webpack": "^5.72.1",
     "wnumb": "^1.2.0",
     "xml2js": "^0.5.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/language-service": "^15.2.8",
+    "@angular-eslint/builder": "16.3.1",
+    "@angular-eslint/eslint-plugin": "16.3.1",
+    "@angular-eslint/eslint-plugin-template": "16.3.1",
+    "@angular-eslint/template-parser": "16.3.1",
+    "@angular/language-service": "^17.0.8",
     "@bartholomej/ngx-translate-extract": "^8.0.2",
-    "@ngrx/schematics": "^14.3.2",
+    "@ngrx/schematics": "^16.3.0",
     "@types/jasmine": "^3.6.11",
     "@types/jasminewd2": "^2.0.8",
     "@types/node": "^12.20.12",
-    "@typescript-eslint/eslint-plugin": "^5.43.0",
-    "@typescript-eslint/parser": "^5.43.0",
-    "eslint": "^8.28.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
+    "esbuild": "0.20.0",
+    "eslint": "^8.39.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsdoc": "^39.4.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
