@@ -34,21 +34,66 @@ export class SearchTypeSelectorComponent implements OnInit, OnDestroy {
 
 
   public searchType: models.SearchType = models.SearchType.DATASET;
+  public searchTypes = models.SearchType;
+  public iconTypes = models.IconType;
   public searchTranslation = models.SearchTypeTranslation;
   public datasets = derivedDatasets;
   public breakpoint$ = this.screenSize.breakpoint$;
   public breakpoints = Breakpoints;
   public isLoggedIn = false;
-  public searchTypes = models.SearchType;
   private subs = new SubSink();
   public isReadMore = true;
+
+  public searchTypeSelectors: models.SearchTypeSelector[] = [
+    {
+      searchType: models.SearchType.DATASET,
+      nameKey:'GEOGRAPHIC',
+      descriptionKeys:['GEOGRAPHIC_SEARCH_ALLOWS_YOU_TO_SEARCH_FOR_DATA_BASED_UPON_A_GEOGRAPHIC_LOCATION'],
+      helpUrl:'https://docs.asf.alaska.edu/vertex/manual/#geographic-search-options',
+      icon:'travel_explore',
+      iconType: models.IconType.MATERIAL,
+    }, {
+      searchType: models.SearchType.LIST,
+      nameKey: 'LIST',
+      descriptionKeys:['LIST_SEARCH_ALLOWS_YOU_TO_ENTER_OR_UPLOAD_A_LIST_OF_SCENE_OR_FILE_NAMES'],
+      helpUrl:'https://docs.asf.alaska.edu/vertex/manual/#list-search-options',
+      icon: 'list',
+      iconType: models.IconType.MATERIAL,
+    }, {
+      searchType: models.SearchType.BASELINE,
+      nameKey:'BASELINE',
+      descriptionKeys:[
+        'BASELINE_SEARCH_PROVIDES_VISUALIZATION_OF_PERPENDICULAR_AND_TEMPORAL_BASELINE_DATA_FOR_A_CHOSEN',
+        'REFERENCE_SCENE',
+      ],
+      helpUrl:'https://docs.asf.alaska.edu/vertex/manual/#baseline-search-options',
+      icon:'assets/icons/baseline-chart.jpg',
+      iconType: models.IconType.IMAGE,
+    }, {
+      searchType: models.SearchType.SBAS,
+      nameKey:'SBAS',
+      descriptionKeys:[
+        'SBAS_SEARCH_PROVIDES_PERPENDICULAR_AND_TEMPORAL_BASELINE_DATA_AS_WELL_AS_SCENE_PAIRS_FOR_A_CHOSEN',
+        'REFERENCE_SCENE',
+      ],
+      helpUrl:'https://docs.asf.alaska.edu/vertex/manual/#sbas-search-options',
+      icon:'assets/icons/sbas-chart.jpg',
+      iconType: models.IconType.IMAGE,
+    }, {
+      searchType: models.SearchType.SARVIEWS_EVENTS,
+      nameKey:'EVENT',
+      descriptionKeys:['EVENT_SEARCH_HARNESSES_THE_CAPABILITIES_OF_SAR_PROCESSING_TO_MONITOR_NATURAL_DISASTERS'],
+      helpUrl:'https://docs.asf.alaska.edu/vertex/manual/#event-search-options',
+      icon:'volcano',
+      iconType: models.IconType.MATERIAL,
+    }
+  ];
 
   constructor(
     public translate: TranslateService,
     private store$: Store<AppState>,
     private screenSize: ScreenSizeService
-  )
-  {}
+  ) {}
 
   ngOnInit() {
     this.subs.add(
@@ -100,14 +145,17 @@ export class SearchTypeSelectorComponent implements OnInit, OnDestroy {
     event.stopPropagation();
   }
 
+  public onCloseMenu(event: Event) {
+    this.trigger.closeMenu();
+    event.stopPropagation();
+  }
+
   public onSearchTypeMenuOpen() {
     const panelId = this.searchMenu.panelId;
     document.getElementById(panelId).focus();
     setTimeout(() => {
       this.searchMenu.focusFirstItem();
       this.searchMenu.resetActiveItem();
-      document.getElementById('firstItem').focus();
-      document.getElementById(panelId).focus();
     }, 10);
   }
 
