@@ -6,8 +6,8 @@ import { MatSharedModule } from '@shared';
 import { MatSelectModule } from '@angular/material/select';
 
 import { SearchTypeSelectorComponent } from './search-type-selector.component';
-import {MatMenuModule} from '@angular/material/menu';
-import {DocsModalModule} from '@components/shared/docs-modal';
+import { MatMenuModule } from '@angular/material/menu';
+import { DocsModalModule } from '@components/shared/docs-modal';
 import { TranslateModule } from "@ngx-translate/core";
 
 @NgModule({
